fix: model selection should update the account status for failurs
