+++ internal/controller/utils/nim.go
@@ -23,7 +23,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"slices"
 	"strings"
 	"time"
 
@@ -591,78 +590,4 @@ func (r *NIMCleanupRunner) Start(ctx context.Context) error {
 // ObjectKeyFromReference returns the ObjectKey given a ObjectReference
 func ObjectKeyFromReference(ref *corev1.ObjectReference) client.ObjectKey {
 	return types.NamespacedName{Name: ref.Name, Namespace: ref.Namespace}
-}
-
-// FilterAvailableNimRuntimes is used for filtering a list of available NIM custom runtimes by a list of model Ids
-func FilterAvailableNimRuntimes(availableRuntimes []NimRuntime, selectedModelList []string,
-	logger logr.Logger) []NimRuntime {
-	// log the available model Ids
-	var size = len(availableRuntimes)
-	modelIds := make([]string, size)
-	for i := 0; i < size; i++ {
-		modelIds[i] = availableRuntimes[i].Name
-	}
-	logger.V(1).Info("fetched the available NIM models",
-		"model Ids", json.RawMessage(`["`+strings.Join(modelIds, `", "`)+`"]`))
-
-	// check the selected models
-	if len(selectedModelList) > 0 {
-		logger.V(1).Info("got the selected NIM model list",
-			"model Ids", json.RawMessage(`["`+strings.Join(selectedModelList, `", "`)+`"]`))
-		// filter the available runtimes
-		var selectedRuntimes []NimRuntime
-		for _, r := range availableRuntimes {
-			if slices.Contains(selectedModelList, r.Name) {
-				selectedRuntimes = append(selectedRuntimes, r)
-			}
-		}
-		return selectedRuntimes
-	}
-
-	return availableRuntimes
-}
-
-// GetSelectedModelList returns a list of Ids of the selected models
-func GetSelectedModelList(
-	ctx context.Context, cmRef *corev1.ObjectReference,
-	namespace string, kubeClient client.Client, logger logr.Logger) ([]string, error) {
-	logger.V(1).Info("getting selected model list")
-
-	// if selected model list is not set
-	if cmRef == nil || len(cmRef.Name) == 0 {
-		return []string{}, nil
-	}
-
-	// get the config map that contains the selected model list
-	cmNs := cmRef.Namespace
-	if cmNs == "" {
-		cmNs = namespace
-	}
-	modelListCm := &corev1.ConfigMap{}
-	modelListCmSubject := types.NamespacedName{Name: cmRef.Name, Namespace: cmNs}
-	if err := kubeClient.Get(ctx, modelListCmSubject, modelListCm); err != nil {
-		if k8serrors.IsNotFound(err) {
-			logger.Error(err, "failed to fetch the config map for getting the selected model list",
-				"config map", modelListCmSubject)
-			return []string{}, nil
-		}
-		return nil, err
-	}
-
-	// convert the selected model list to a string array
-	if modelListCm.Data != nil {
-		if models, ok := modelListCm.Data["models"]; ok {
-			var selectedModelIds []string
-			if err := json.Unmarshal([]byte(models), &selectedModelIds); err != nil {
-				logger.Error(err, "failed to unmarshal the selected mode list",
-					"model list", models)
-				return []string{}, nil
-			}
-			return selectedModelIds, nil
-		}
-	}
-
-	logger.Error(nil, "failed to get the selected model list from the data of the config map",
-		"config map", modelListCmSubject)
-	return []string{}, nil
 }
